added devcontainer, progress on VOMS support
Added the devcontainer, which allows the user to
develop the application using the enviroment
of a specified container in devcontainer
Progress was made in VOMS support, you can find some of the commands used
inside the debug requester.
Also, some lines on the nginx.conf were commented, and will remain as such
until further research.