Enables VOMS certificates support
A lot of fixes have been made. Added some readmes to better explain the behaviour
of each component in this repository.
The user certificate is now formatted correctly, and can be used to make a server request using VOMS
The docker-compose file has been updated, so that it installs all the packages
needed to validate the AC.
Finally, the debugger made with curl has also been fixed.